Honor PLATLIBDIR during install
Python builds with an expectation that lib-dynload exists in $(PLATLIBDIR)/python3.9/lib-dynload.
platlibdir defaults to lib.

During the install process, libdynload is copied to $(BINLIBDEST)/lib-dynload, which is based
off of LIBDIR rather than PLATLIBDIR.

This succeeds as long as either PLATLIBDIR is specified to match LIBDIR, or LIBDIR is also lib (common)